refactor(challengeId): Move components out from 'components' folder
