+++ apps/nova/src/app/[locale]/(dashboard)/challenges/[challengeId]/countdown-timer.tsx
