+++ apps/nova/src/app/[locale]/(dashboard)/challenges/[challengeId]/prompt-form.tsx
